unistd.h more architecture independent without *long32_t.
There is some problems with the data types and its sizes across
different platforms. The actual problem here resides with "long",
which have 4 bytes in x86 and 8 bytes in x86_64.

The "long"s typedefs were removed from unistd.h because they can't be
coherent across our platforms but its behavior is a bless when we talk
about "size_t" because it reflects exactly what we intent with it.

The sources that were using "ulong32_t" and "long32_t" were changed to
"uint32_t" and "int32_t".